# specs/api.openapi.yaml
components:
  schemas:
    AgentCreateRequest:
      description: Request to create a new agent.
      properties:
        abuse_config:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Per-agent abuse settings overrides (keys match AbuseConfig
            fields)
          title: Abuse Config
        description:
          default: ''
          description: Human-readable description
          title: Description
          type: string
        llm_config:
          anyOf:
          - $ref: '#/components/schemas/LLMConfig'
          - type: 'null'
          description: Per-agent LLM overrides
        name:
          description: Unique agent name (lowercase, no spaces)
          minLength: 1
          pattern: ^[a-z][a-z0-9_-]*$
          title: Name
          type: string
        provider_priority:
          description: Provider names in priority order (from ProviderStore)
          items:
            type: string
          title: Provider Priority
          type: array
        system_prompt:
          anyOf:
          - type: string
          - type: 'null'
          description: Per-agent system prompt override
          title: System Prompt
        tenant_ids:
          description: Tenant IDs for this agent
          items:
            type: string
          title: Tenant Ids
          type: array
        voice_config:
          anyOf:
          - $ref: '#/components/schemas/VoiceAgentConfig'
          - type: 'null'
          description: Per-agent voice configuration overrides
        widget_config:
          anyOf:
          - $ref: '#/components/schemas/WidgetConfig'
          - type: 'null'
          description: Embed widget display configuration
      required:
      - name
      title: AgentCreateRequest
      type: object
    AgentListResponse:
      description: List of agents.
      properties:
        agents:
          description: List of agents
          items:
            $ref: '#/components/schemas/AgentResponse'
          title: Agents
          type: array
      required:
      - agents
      title: AgentListResponse
      type: object
    AgentResponse:
      description: Agent metadata.
      properties:
        abuse_config:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Per-agent abuse settings overrides
          title: Abuse Config
        created_at:
          description: ISO timestamp
          title: Created At
          type: string
        description:
          default: ''
          description: Human-readable description
          title: Description
          type: string
        llm_config:
          anyOf:
          - $ref: '#/components/schemas/LLMConfig'
          - type: 'null'
          description: Per-agent LLM overrides
        name:
          description: Unique agent name
          title: Name
          type: string
        provider_priority:
          description: Provider names in priority order (from ProviderStore)
          items:
            type: string
          title: Provider Priority
          type: array
        system_prompt:
          anyOf:
          - type: string
          - type: 'null'
          description: Per-agent system prompt override
          title: System Prompt
        tenant_ids:
          description: Tenant IDs
          items:
            type: string
          title: Tenant Ids
          type: array
        updated_at:
          description: ISO timestamp
          title: Updated At
          type: string
        voice_config:
          anyOf:
          - $ref: '#/components/schemas/VoiceAgentConfig'
          - type: 'null'
          description: Per-agent voice configuration overrides
        widget_config:
          anyOf:
          - $ref: '#/components/schemas/WidgetConfig'
          - type: 'null'
          description: Embed widget display configuration
      required:
      - name
      - created_at
      - updated_at
      title: AgentResponse
      type: object
    AgentUpdateRequest:
      description: Request to update an existing agent.
      properties:
        abuse_config:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Per-agent abuse settings overrides
          title: Abuse Config
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Human-readable description
          title: Description
        llm_config:
          anyOf:
          - $ref: '#/components/schemas/LLMConfig'
          - type: 'null'
          description: Per-agent LLM overrides
        provider_priority:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          description: Provider names in priority order (from ProviderStore)
          title: Provider Priority
        system_prompt:
          anyOf:
          - type: string
          - type: 'null'
          description: Per-agent system prompt override
          title: System Prompt
        tenant_ids:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          description: Tenant IDs for this agent
          title: Tenant Ids
        voice_config:
          anyOf:
          - $ref: '#/components/schemas/VoiceAgentConfig'
          - type: 'null'
          description: Per-agent voice configuration overrides
        widget_config:
          anyOf:
          - $ref: '#/components/schemas/WidgetConfig'
          - type: 'null'
          description: Embed widget display configuration
      title: AgentUpdateRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    LLMConfig:
      description: "Per-agent LLM configuration. All fields optional \u2014 falls\
        \ back to global defaults when null."
      properties:
        api_base:
          anyOf:
          - type: string
          - type: 'null'
          description: Base URL for the API
          title: Api Base
        api_key:
          anyOf:
          - type: string
          - type: 'null'
          description: API key for the provider
          title: Api Key
        max_tokens:
          anyOf:
          - minimum: 1.0
            type: integer
          - type: 'null'
          description: Maximum tokens in response
          title: Max Tokens
        model:
          anyOf:
          - type: string
          - type: 'null'
          description: Model name (e.g. qwen2.5:0.5b or mistral/mistral-small)
          title: Model
        provider:
          anyOf:
          - type: string
          - type: 'null'
          description: 'LLM provider: ollama, mistral, openai, anthropic'
          title: Provider
        system_prompt:
          anyOf:
          - type: string
          - type: 'null'
          description: System prompt (overrides global)
          title: System Prompt
        temperature:
          anyOf:
          - maximum: 2.0
            minimum: 0.0
            type: number
          - type: 'null'
          description: Model temperature
          title: Temperature
      title: LLMConfig
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    VoiceAgentConfig:
      description: 'Per-agent STT voice configuration overrides.


        All fields are optional (None = fall back to global VoiceConfig).'
      properties:
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Override master switch for this agent
          title: Enabled
        stt_fallback:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Override STT fallback setting for this agent
          title: Stt Fallback
        stt_provider:
          anyOf:
          - type: string
          - type: 'null'
          description: STT provider name (must match a name in VoiceConfig.stt_providers)
          title: Stt Provider
        voice_input_disabled:
          anyOf:
          - type: boolean
          - type: 'null'
          description: Explicitly disable voice input for this agent (hide mic in
            widget)
          title: Voice Input Disabled
      title: VoiceAgentConfig
      type: object
    WidgetConfig:
      description: Embed widget display configuration for an agent.
      properties:
        accent_color:
          default: '#0f766e'
          description: Accent color (hex)
          title: Accent Color
          type: string
        greeting:
          default: "\u0427\u0435\u043C \u043C\u043E\u0433\u0443 \u043F\u043E\u043C\
            \u043E\u0447\u044C?"
          description: Greeting message in chat
          title: Greeting
          type: string
        position:
          default: right
          description: 'Widget position: right | left'
          title: Position
          type: string
        title:
          default: "\u0410\u0441\u0441\u0438\u0441\u0442\u0435\u043D\u0442"
          description: Widget header title
          title: Title
          type: string
      title: WidgetConfig
      type: object
info:
  description: LLM agent orchestration service
  title: Helperium API
  version: 1.1.0
openapi: 3.1.0
paths:
  /admin/abuse-config:
    get:
      description: Get the current effective abuse configuration (from file + env).
      operationId: get_abuse_config_admin_abuse_config_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Get Abuse Config
    post:
      description: Save new abuse config directly (admin dashboard alternative endpoint).
      operationId: save_abuse_config_admin_abuse_config_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              title: Data
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Save Abuse Config
...
  /admin/guardrails:
    get:
      description: Get current guard config.
      operationId: get_guardrails_admin_guardrails_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Get Guardrails
    post:
      description: 'Update guard config. Accepts: enabled, block_on_match.'
      operationId: update_guardrails_admin_guardrails_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              title: Config
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Guardrails
...
  /admin/llm-providers:
    get:
      description: List all LLM providers with masked API keys.
      operationId: list_llm_providers_admin_llm_providers_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: List Llm Providers
    post:
      description: "Add a new LLM provider.\n\nBody:\n    name (required): unique\
        \ provider name\n    model (required): model identifier (e.g. openai/gpt-4o,\
        \ anthropic/claude-3-sonnet)\n    provider: provider type \u2014 auto-detected\
        \ from model prefix if omitted\n    api_key: API key (will not be returned\
        \ in full)\n    api_base: custom API base URL\n    enabled: whether the provider\
        \ is active (default: true)"
      operationId: add_llm_provider_admin_llm_providers_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              title: Body
              type: object
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add Llm Provider
  /admin/llm-providers/{name}:
    delete:
      description: Delete an LLM provider.
      operationId: delete_llm_provider_admin_llm_providers__name__delete
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Llm Provider
    get:
      description: Get a single LLM provider with masked API key.
      operationId: get_llm_provider_admin_llm_providers__name__get
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Llm Provider
    put:
      description: 'Update an existing LLM provider.


        Omitted fields keep their current value.

        Set api_key="" to keep existing key (not change it).

        Set api_key="__clear__" to clear the key.'
      operationId: update_llm_provider_admin_llm_providers__name__put
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              title: Body
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Llm Provider
  /admin/llm-providers/{name}/toggle:
    post:
      description: Toggle a provider on/off.
      operationId: toggle_llm_provider_admin_llm_providers__name__toggle_post
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Toggle Llm Provider
...
  /admin/spending/{tenant_id}:
    get:
      description: Get spending for a specific tenant.
      operationId: get_tenant_spending_admin_spending__tenant_id__get
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Tenant Spending
    post:
      description: 'Set per-tenant budget override.


        Body: {"budget": 100.0}'
      operationId: set_tenant_budget_admin_spending__tenant_id__post
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              title: Config
              type: object
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Set Tenant Budget
  /api/agents:
    get:
      operationId: list_agents_endpoint_api_agents_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
          description: Successful Response
      summary: List Agents Endpoint
    post:
      operationId: create_agent_endpoint_api_agents_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Agent Endpoint
  /api/agents/{name}:
    delete:
      operationId: delete_agent_endpoint_api_agents__name__delete
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Agent Endpoint
    get:
      operationId: get_agent_endpoint_api_agents__name__get
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Agent Endpoint
    put:
      operationId: update_agent_endpoint_api_agents__name__put
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdateRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update Agent Endpoint
  /api/agents/{name}/widget-config:
    get:
      operationId: agent_widget_config_endpoint_api_agents__name__widget_config_get
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Agent Widget Config Endpoint Api Agents  Name  Widget
                  Config Get
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Agent Widget Config Endpoint
...